Updated field type

+++ Masya.TelegramBot.DataAccess/Models/Favorites.cs
@@ -9,7 +9,7 @@ namespace Masya.TelegramBot.DataAccess.Models
         [Key]
         public int Id { get; set; }
 
-        public int? UserId { get; set; }
+        public long? UserId { get; set; }
         public User User { get; set; }
 
         public int? RealtyObjectId { get; set; }
